Drastic changes
New archiheture for exceptions:
1.Now there is only one class that extends RunTime exception for each entity
2.They do not have the last annotation
3.They accept messages from enum which:
-Keep all of the possible errors that could be throw
4.P/C:
+More clear code
+Easy to see all of the possible errors at a place
+No magic strings
- No specific status code (This will be changed at later time ,when i finish with testing different approaches)
------>
Unit testing for Services ,but its really bad, because currently Im trying to find the best archiheture  to structure a unit test in bigger project and to be future proof so:
1.Every test is isolated (Doesnt depend on global variables)
2.JUnit and Mockito are used
3.There is a lot of work on them ,but currently this are my "bigger" unit tests that I have write to now ,cuz the last were on only one method with hundred of different outputs ,when here isn't like there

+++ back-end/src/main/java/ichop/service/user/UserServicesImp.java
@@ -7,18 +7,15 @@ import ichop.domain.entities.users.UserRole;
 import ichop.domain.models.binding.user.UserRegisterBindingModel;
 import ichop.domain.models.binding.user.UserForgottenPasswordBindingModel;
 import ichop.domain.models.binding.user.UserResetPasswordBindingModel;
-import ichop.exceptions.token.TokenCannotBeNullException;
-import ichop.exceptions.token.TokenNotValidException;
-import ichop.exceptions.user.PasswordsDoesntMatchException;
-import ichop.exceptions.user.UserAlreadyExistsException;
-import ichop.exceptions.user.UserCannotBeNullException;
-import ichop.repository.user.UserRepository;
+import ichop.exceptions.token.TokenException;
+import ichop.exceptions.token.TokenExceptionMessages;
+import ichop.exceptions.user.UserException;
+import ichop.exceptions.user.UserExceptionMessages;
 import ichop.service.role.UserRoleServices;
 import ichop.service.role.UserRoles;
 import ichop.service.token.PasswordResetTokenServices;
 import org.modelmapper.ModelMapper;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.GrantedAuthority;
+import org.springframework.http.HttpStatus;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
@@ -27,75 +24,63 @@ import org.springframework.stereotype.Service;
 import java.time.LocalDateTime;
 import java.util.HashSet;
 import java.util.Set;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 @Service
 public class UserServicesImp implements UserServices {
 
-    private static final String EMAIL_PATTERN = "(?:[a-z0-9!#$%&'*+\\=?^_`{|}~-]+(?:\\.[a-z0-9!#$%&'*+\\=?^_`{|}~-]+)*|\"(?:[\\x01-\\x08\\x0b\\x0c\\x0e-\\x1f\\x21\\x23-\\x5b\\x5d-\\x7f]|\\\\[\\x01-\\x09\\x0b\\x0c\\x0e-\\x7f])*\")@(?:(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\\.)+[a-z0-9](?:[a-z0-9-]*[a-z0-9])?|\\[(?:(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)\\.){3}(?:25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?|[a-z0-9-]*[a-z0-9]:(?:[\\x01-\\x08\\x0b\\x0c\\x0e-\\x1f\\x21-\\x5a\\x53-\\x7f]|\\\\[\\x01-\\x09\\x0b\\x0c\\x0e-\\x7f])+)\\])";
-
-    private final UserRepository userRepository;
-
     private final UserRoleServices userRoleServices;
+    private final UserBaseServices userBaseServices;
+    private final EmailServices emailServices;
     private final PasswordResetTokenServices passwordResetTokenServices;
 
     private final ModelMapper modelMapper;
     private final BCryptPasswordEncoder passwordEncoder;
-    private final EmailServices emailServices;
 
-    @Autowired
-    public UserServicesImp(UserRepository userRepository, UserRoleServices userRoleServices, PasswordResetTokenServices passwordResetTokenServices, ModelMapper modelMapper, BCryptPasswordEncoder passwordEncoder, EmailServices emailServices) {
-        this.userRepository = userRepository;
+    public UserServicesImp(UserRoleServices userRoleServices, UserBaseServices userBaseServices, EmailServices emailServices, PasswordResetTokenServices passwordResetTokenServices, ModelMapper modelMapper, BCryptPasswordEncoder passwordEncoder) {
         this.userRoleServices = userRoleServices;
+        this.userBaseServices = userBaseServices;
+        this.emailServices = emailServices;
         this.passwordResetTokenServices = passwordResetTokenServices;
         this.modelMapper = modelMapper;
         this.passwordEncoder = passwordEncoder;
-        this.emailServices = emailServices;
     }
+
 
     @Override
     public UserDetails loadUserByUsername(String usernameOrEmail) throws UsernameNotFoundException {
-        if (this.isEmail(usernameOrEmail)) {
+        if (this.userBaseServices.isEmail(usernameOrEmail)) {
 
-            UserDetails userDetails = this.userRepository.findUserByEmail(usernameOrEmail);
+            UserDetails userDetails = this.userBaseServices.getUserByEmail(usernameOrEmail);
 
             if(userDetails != null){
                 return userDetails;
             }
 
-            throw new UsernameNotFoundException("Sorry ,but user with the provided email has not been found!");
+            throw new UsernameNotFoundException(UserExceptionMessages.USER_WITH_THAT_USERNAME_NOT_FOUND.getDescription());
         }
 
-        UserDetails userDetails = this.userRepository.findUserByUsername(usernameOrEmail);
+        UserDetails userDetails = this.userBaseServices.getUserByUsername(usernameOrEmail);
 
         if(userDetails != null){
             return userDetails;
         }
 
-        throw new UsernameNotFoundException("Sorry ,but user with the provided username has not been found!");
-    }
-
-    @Override
-    public boolean isEmail(String value) {
-        Matcher matcher = Pattern.compile(EMAIL_PATTERN)
-                .matcher(value);
-        return matcher.find();
+        throw new UsernameNotFoundException(UserExceptionMessages.USER_WITH_THAT_EMAIL_NOT_FOUND.getDescription());
     }
 
     @Override
     public User register(UserRegisterBindingModel userRegisterBindingModel) {
 
-        if (this.existsByUsername(userRegisterBindingModel.getUsername())) {
-            throw new UserAlreadyExistsException();
+        if (this.userBaseServices.existsByUsername(userRegisterBindingModel.getUsername())) {
+            throw new UserException(UserExceptionMessages.USERNAME_ALREADY_EXISTS);
         }
 
-        if (this.existsByEmail(userRegisterBindingModel.getEmail())) {
-            throw new UserAlreadyExistsException();
+        if (this.userBaseServices.existsByEmail(userRegisterBindingModel.getEmail())) {
+            throw new UserException(UserExceptionMessages.EMAIL_ALREADY_EXISTS);
         }
 
         if(!userRegisterBindingModel.getPassword().equals(userRegisterBindingModel.getConfirmPassword())){
-            throw new PasswordsDoesntMatchException();
+            throw new UserException(UserExceptionMessages.PASSWORDS_DOESNT_MATCH);
         }
 
         User user = this.modelMapper.map(userRegisterBindingModel, User.class);
@@ -107,29 +92,15 @@ public class UserServicesImp implements UserServices {
 
         user.setAuthorities(this.getInitialAuthorities());
 
-        this.userRepository.save(user);
+        this.userBaseServices.save(user);
 
         return user;
-    }
-
-    @Override
-    public boolean existsByUsername(String username) {
-        return this.userRepository.findUserByUsername(username) != null;
-    }
-
-    @Override
-    public boolean existsByEmail(String email) {
-        return this.userRepository.findUserByEmail(email) != null;
-    }
-
-    private Integer getTotalUsers() {
-        return this.userRepository.findAll().size();
     }
 
     private Set<UserRole> getInitialAuthorities() {
         Set<UserRole> userRoles = new HashSet<>();
 
-        if (this.getTotalUsers() == 0) {
+        if (this.userBaseServices.getTotalUsers() == 0) {
             userRoles.add(this.userRoleServices.create(UserRoles.OWNER));
             userRoles.add(this.userRoleServices.create(UserRoles.ADMIN));
             userRoles.add(this.userRoleServices.create(UserRoles.MODERATOR));
@@ -141,30 +112,6 @@ public class UserServicesImp implements UserServices {
     }
 
     @Override
-    public UserRole getRole(User user) {
-
-        if (user == null) {
-            throw new UserCannotBeNullException();
-        }
-
-        return (UserRole) user
-                .getAuthorities()
-                .stream()
-                .min((x1, x2) -> Integer.compare(UserRoles.valueOf(((GrantedAuthority) x2).getAuthority()).ordinal(), UserRoles.valueOf(((GrantedAuthority) x1).getAuthority()).ordinal()))
-                .orElse(null);
-    }
-
-    @Override
-    public User getUserByUsername(String username) {
-        return this.userRepository.findUserByUsername(username);
-    }
-
-    @Override
-    public User getUserByEmail(String email) {
-        return this.userRepository.findUserByEmail(email);
-    }
-
-    @Override
     public void sendPasswordResetEmail(UserForgottenPasswordBindingModel userForgottenPasswordBindingModel) {
 
         User user = (User) this.loadUserByUsername(userForgottenPasswordBindingModel.getUsernameOrEmail());
@@ -172,7 +119,7 @@ public class UserServicesImp implements UserServices {
         PasswordResetToken passwordResetToken = this.passwordResetTokenServices.createToken(user);
 
         if (passwordResetToken == null) {
-            throw new TokenCannotBeNullException();
+            throw new TokenException(TokenExceptionMessages.TOKEN_IS_NULL);
         }
 
         this.emailServices.sendResetPasswordEmail(user.getEmail(),passwordResetToken.getToken(),passwordResetToken.getExpiryDate());
@@ -182,14 +129,18 @@ public class UserServicesImp implements UserServices {
     public void resetPassword(UserResetPasswordBindingModel userResetPasswordBindingModel, String resetToken) {
 
         if (!this.passwordResetTokenServices.isValid(resetToken)) {
-            throw new TokenNotValidException();
+            throw new TokenException(TokenExceptionMessages.TOKEN_IS_NOT_VALID);
+        }
+
+        if(!userResetPasswordBindingModel.getPassword().equals(userResetPasswordBindingModel.getConfirmPassword())){
+            throw new UserException(UserExceptionMessages.PASSWORDS_DOESNT_MATCH);
         }
 
         User user = this.passwordResetTokenServices.getTokenByToken(resetToken).getUser();
 
         user.setPassword(this.passwordEncoder.encode(userResetPasswordBindingModel.getPassword()));
 
-        this.userRepository.save(user);
+        this.userBaseServices.save(user);
         this.passwordResetTokenServices.deleteOldestToken(user);
     }
 }
